undo rename which was just a test of LLM-based rename
